scripts/requirements-verify.txt
fastapi
pydantic
pydantic-settings
python-dotenv
azure-identity
azure-core
httpx
python-jose[cryptography]
python-multipart
websockets
